+++ transport_udp_test.go
@@ -18,11 +18,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/uber/jaeger-client-go/internal/transport"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/uber/jaeger-client-go/internal/reporterstats"
 	"github.com/uber/jaeger-client-go/testutils"
 	"github.com/uber/jaeger-client-go/thrift"
 	j "github.com/uber/jaeger-client-go/thrift-gen/jaeger"
@@ -33,7 +32,7 @@ var (
 	jaegerTracer  = testTracer.(*Tracer)
 
 	// API check
-	_ transport.ReporterStatsReceiver = new(udpSender)
+	_ reporterstats.Receiver = new(udpSender)
 )
 
 func getThriftSpanByteLength(t *testing.T, span *Span) int {
